WIP CPU-only LM PyO3, fix CPU gen bug for dual_ar
