+++ Cargo.toml
@@ -17,3 +17,4 @@ candle-core = "0.8.3"
 candle-nn = "0.8.3"
 candle-transformers = "0.8.3"
 candle-flash-attn = "0.8.3"
+tokenizers = "0.20.0"
