+++ fish_speech_core/lib/lm/mod.rs
@@ -3,4 +3,5 @@ pub mod generate;
 mod ops;
 pub mod sampling;
 
+pub use dual_ar::BaseModelArgs;
 pub use dual_ar::DualARTransformer;
